added window for creatures
