+++ Initiative/InitiativeWindow.py
@@ -1,3 +1,4 @@
+from Creatures.CreatureWindow import CreatureWindow
 import tkinter.ttk as ttk
 from Initiative.InitiativeHandler import InitiativeHandler
 
@@ -15,6 +16,8 @@ class InitiativeWindow(ttk.Frame):
         ttk.Button(self, text="Next turn", 
             command=self.nextTurn).grid(column=1, row=1)
 
+        self.currentCreatureWindow = ttk.Frame(self)
+
     def startInitiative(self):
         self.initiativeHandler.StartInitiative()
 
@@ -24,8 +27,16 @@ class InitiativeWindow(ttk.Frame):
     def handleInitiativeUpdate(self, data):
         self.initiativeList.grid_forget()
         self.initiativeList = ttk.Frame(self)
-        self.initiativeList.grid(column=0, row=0)
+        self.initiativeList.grid(column=0, row=0, rowspan=2)
 
         for (controller, creature), initiative in data:
             ttk.Label(master=self.initiativeList, 
                 text=f"{creature}: {controller}: {initiative}").pack()
+
+    def handleStartTurn(self, data):
+        print(data)
+
+        self.currentCreatureWindow.grid_forget()
+
+        self.currentCreatureWindow = CreatureWindow(data, self.initiativeHandler, self)
+        self.currentCreatureWindow.grid(column=2, row=0, rowspan=2)
